Revert "[rocjitsu] DBI WIP: Added end-to-end tests for nop probe"
This reverts commit a940a598cd21261df2b916ec493dcb93b6fc0a98.
